First install settings vs new features settings
This commit makes it so that on first install, all settings are set to their defaults. Should the extension be updated with a new feature after this, new features will by default set to "false" as opposed to set to their default option.

Should anyone wish this changed, feel free to make an issue stating so.